Creating the views
The main view for the calculator app will be index, that is contained in
the maths direcotry.
Also, the shared dir, there are two files, the _calculator.html.erb file
contain the containers for the calculator.
The _basic_mode_calculator.html.erb contains the strtuctor for the
calculator, that is, the keyboard and the display, and it will be
rendered in the _calculator.html.erb.