Styling the views
The app will be styled using SASS, for that reason it was renamed the application.css to application.scss.
The app/assets/stylesheets/calculator directory contiains the files for the whole
calculator styles.
The app/assets/stylesheets/default/ directory contains the basic styles
for the app, that is reset default styles for the browser, set the main
background color, etc. Also contains the file for the scss variables.

And finally, the mahts.scss fiel contains only the imports for the
necessary files for the calculator.